+++ packages/atomic/fake-loader/atomic.esm.js
@@ -1,0 +1,6 @@
+import {defineCustomElements, atomicVersion, headlessVersion} from './index.js';
+
+export {atomicVersion, headlessVersion};
+
+window.CoveoAtomic = {version: atomicVersion, headlessVersion: headlessVersion};
+defineCustomElements();