+++ packages/atomic/src/components/commerce/atomic-commerce-recommendation-list/atomic-commerce-recommendation-list.tsx
@@ -24,10 +24,9 @@ import {
 import {FocusTargetController} from '../../../utils/stencil-accessibility-utils';
 import {randomID} from '../../../utils/stencil-utils';
 import {ResultsPlaceholdersGuard} from '../../common/atomic-result-placeholder/stencil-placeholders';
-import {Carousel} from '../../common/carousel';
 import {createAppLoadedListener} from '../../common/interface/store';
-import {DisplayGrid} from '../../common/item-list/display-grid';
 import {ItemDisplayGuard} from '../../common/item-list/item-display-guard';
+import {DisplayGrid} from '../../common/item-list/stencil-display-grid';
 import {DisplayWrapper} from '../../common/item-list/stencil-display-wrapper';
 import {
   ItemListCommon,
@@ -39,6 +38,7 @@ import {
   ItemDisplayImageSize,
   getItemListDisplayClasses,
 } from '../../common/layout/display-options';
+import {Carousel} from '../../common/stencil-carousel';
 import {Heading} from '../../common/stencil-heading';
 import {Hidden} from '../../common/stencil-hidden';
 import {CommerceBindings} from '../atomic-commerce-recommendation-interface/atomic-commerce-recommendation-interface';