+++ packages/atomic/src/components/insight/result-templates/atomic-insight-result-template/atomic-insight-result-template.tsx
@@ -18,7 +18,7 @@ import {
   shadow: true,
 })
 export class AtomicInsightResultTemplate {
-  private resultTemplateCommon: ResultTemplateCommon;
+  private resultTemplateCommon!: ResultTemplateCommon;
 
   @State() public error!: Error;
 
@@ -68,7 +68,9 @@ export class AtomicInsightResultTemplate {
     string[]
   > = {};
 
-  constructor() {
+  constructor() {}
+
+  connectedCallback() {
     this.resultTemplateCommon = new ResultTemplateCommon({
       host: this.host,
       setError: (err) => {