+++ packages/atomic/src/components/ipx/atomic-ipx-recs-list/atomic-recs-list/atomic-ipx-recs-list.tsx
@@ -29,11 +29,10 @@ import {
 import {FocusTargetController} from '../../../../utils/stencil-accessibility-utils';
 import {randomID} from '../../../../utils/stencil-utils';
 import {ResultsPlaceholdersGuard} from '../../../common/atomic-result-placeholder/stencil-placeholders';
-import {Carousel} from '../../../common/carousel';
 import {createAppLoadedListener} from '../../../common/interface/store';
-import {DisplayGrid} from '../../../common/item-list/display-grid';
 import {ItemDisplayGuard} from '../../../common/item-list/item-display-guard';
 import {ItemTemplateProvider} from '../../../common/item-list/item-template-provider';
+import {DisplayGrid} from '../../../common/item-list/stencil-display-grid';
 import {DisplayWrapper} from '../../../common/item-list/stencil-display-wrapper';
 import {
   ItemListCommon,
@@ -45,6 +44,7 @@ import {
   ItemDisplayDensity,
   ItemDisplayImageSize,
 } from '../../../common/layout/display-options';
+import {Carousel} from '../../../common/stencil-carousel';
 import {Heading} from '../../../common/stencil-heading';
 import {RecsBindings} from '../../../recommendations/atomic-recs-interface/atomic-recs-interface';