+++ packages/atomic/src/components/recommendations/atomic-recs-list/atomic-recs-list.tsx
@@ -25,11 +25,10 @@ import {
 import {FocusTargetController} from '../../../utils/stencil-accessibility-utils';
 import {randomID} from '../../../utils/stencil-utils';
 import {ResultsPlaceholdersGuard} from '../../common/atomic-result-placeholder/stencil-placeholders';
-import {Carousel} from '../../common/carousel';
 import {createAppLoadedListener} from '../../common/interface/store';
-import {DisplayGrid} from '../../common/item-list/display-grid';
 import {ItemDisplayGuard} from '../../common/item-list/item-display-guard';
 import {ItemTemplateProvider} from '../../common/item-list/item-template-provider';
+import {DisplayGrid} from '../../common/item-list/stencil-display-grid';
 import {DisplayWrapper} from '../../common/item-list/stencil-display-wrapper';
 import {
   ItemListCommon,
@@ -41,6 +40,7 @@ import {
   ItemDisplayBasicLayout,
   getItemListDisplayClasses,
 } from '../../common/layout/display-options';
+import {Carousel} from '../../common/stencil-carousel';
 import {Heading} from '../../common/stencil-heading';
 import {RecsBindings} from '../atomic-recs-interface/atomic-recs-interface';