+++ packages/atomic/src/components/search/facets/atomic-facet/atomic-facet.tsx
@@ -38,7 +38,7 @@ import {
 } from '../../../../utils/stencil-accessibility-utils';
 import {parseDependsOn} from '../../../common/facets/depends-on';
 import {FacetInfo} from '../../../common/facets/facet-common-store';
-import {FacetContainer} from '../../../common/facets/facet-container/facet-container';
+import {FacetContainer} from '../../../common/facets/facet-container/stencil-facet-container';
 import {FacetGuard} from '../../../common/facets/facet-guard';
 import {FacetHeader} from '../../../common/facets/facet-header/stencil-facet-header';
 import {FacetPlaceholder} from '../../../common/facets/facet-placeholder/facet-placeholder';
@@ -51,7 +51,7 @@ import {
 } from '../../../common/facets/facet-search/facet-search-utils';
 import {FacetSearchMatches} from '../../../common/facets/facet-search/stencil-facet-search-matches';
 import {FacetSearchValue} from '../../../common/facets/facet-search/stencil-facet-search-value';
-import {FacetShowMoreLess} from '../../../common/facets/facet-show-more-less/facet-show-more-less';
+import {FacetShowMoreLess} from '../../../common/facets/facet-show-more-less/stencil-facet-show-more-less';
 import {
   FacetValueProps,
   FacetValue,