+++ packages/atomic/src/utils/functional-component-utils.ts
@@ -2,7 +2,7 @@ import {nothing, TemplateResult} from 'lit';
 import {DirectiveResult} from 'lit/directive.js';
 
 export interface FunctionalComponent<T> {
-  ({props}: {props: T}): TemplateResult | TemplateResult[] | typeof nothing;
+  ({props}: {props: T}): TemplateResult | typeof nothing;
 }
 
 export interface FunctionalComponentWithChildren<T> {
@@ -11,6 +11,10 @@ export interface FunctionalComponentWithChildren<T> {
   }: {
     props: T;
   }): (children: FunctionalComponentChildren) => TemplateResult;
+}
+
+export interface FunctionalComponentWithChildrenNoProps {
+  (): (children: FunctionalComponentChildren) => TemplateResult;
 }
 
 export interface FunctionalComponentGuard<T> {