+++ packages/atomic/src/utils/resource-url.ts
@@ -9,7 +9,7 @@ function isCoveoCDN() {
 }
 
 function getCoveoCdnResourceUrl() {
-  return import.meta.resolve('../../');
+  return import.meta.resolve('./');
 }
 
 export const getResourceUrl = () =>