+++ packages/atomic/stencil-proxy/esm/loader.js
@@ -1,22 +1,12 @@
-import commerceComponents from '../atomic/components/components/commerce/lazy-index.js';
-import commonComponents from '../atomic/components/components/common/lazy-index.js';
-import insightComponents from '../atomic/components/components/insight/lazy-index.js';
-import ipxComponents from '../atomic/components/components/ipx/lazy-index.js';
-import recommendationsComponents from '../atomic/components/components/recommendations/lazy-index.js';
-import searchComponents from '../atomic/components/components/search/lazy-index.js';
-import {defineCustomElements as originalDefineCustomElements} from './_loader.js';
+import {registerAutoloader} from '../atomic/autoloader/index.esm.js';
+import {defineCustomElements as stencilDefineCustomElements} from './_loader.js';
 
-const defineCustomElements = function (...args) {
-  Object.values({
-    ...commonComponents,
-    ...recommendationsComponents,
-    ...ipxComponents,
-    ...insightComponents,
-    ...commerceComponents,
-    ...searchComponents,
-  }).forEach((importFunction) => importFunction());
-  originalDefineCustomElements(...args);
+const defineCustomElements = async function (...args) {
+  const rootElementAutoloader = args.length > 2 ? args.pop() : undefined;
+  registerAutoloader(rootElementAutoloader);
+  stencilDefineCustomElements(...args);
 };
 
 export * from './_loader.js';
+export * from './version.js';
 export {defineCustomElements};