+++ packages/headless/src/features/commerce/common/base-commerce-api-request-builder.ts
@@ -30,7 +30,7 @@ export const buildBaseCommerceAPIRequest = (
         state.configuration.environment
       ),
     organizationId: state.configuration.organizationId,
-    trackingId: state.configuration.analytics.trackingId,
+    trackingId: state.configuration.analytics.trackingId!,
     ...restOfContext,
     ...(state.configuration.analytics.enabled
       ? {clientId: navigatorContext.clientId}