+++ packages/atomic/CHANGELOG.md
@@ -1,3 +1,49 @@
+## 3.25.0 (2025-05-28)
+
+- chore: add facet-container functional component (#5309) ([41f1eb2](https://github.com/coveo/ui-kit/commits/41f1eb2)), closes [#5309](https://github.com/coveo/ui-kit/issues/5309)
+- chore: migrate facet-common (#5325) ([2807727](https://github.com/coveo/ui-kit/commits/2807727)), closes [#5325](https://github.com/coveo/ui-kit/issues/5325)
+- chore: migrate facet-show-more-less (#5220) ([effd0e4](https://github.com/coveo/ui-kit/commits/effd0e4)), closes [#5220](https://github.com/coveo/ui-kit/issues/5220)
+- chore: migrate facet-value-exclude (#5326) ([200b045](https://github.com/coveo/ui-kit/commits/200b045)), closes [#5326](https://github.com/coveo/ui-kit/issues/5326)
+- chore: skip flaky test in cypress search box (#5364) ([c0efe50](https://github.com/coveo/ui-kit/commits/c0efe50)), closes [#5364](https://github.com/coveo/ui-kit/issues/5364)
+- chore(atomic): migrate breadcrumb-button.tsx (#5350) ([6cb2876](https://github.com/coveo/ui-kit/commits/6cb2876)), closes [#5350](https://github.com/coveo/ui-kit/issues/5350)
+- chore(atomic): migrate breadcrumb-container.tsx (#5352) ([e4ba2aa](https://github.com/coveo/ui-kit/commits/e4ba2aa)), closes [#5352](https://github.com/coveo/ui-kit/issues/5352)
+- chore(atomic): migrate carousel (#5343) ([befd28e](https://github.com/coveo/ui-kit/commits/befd28e)), closes [#5343](https://github.com/coveo/ui-kit/issues/5343)
+- chore(atomic): migrate display-table.tsx (#5274) ([9debf19](https://github.com/coveo/ui-kit/commits/9debf19)), closes [#5274](https://github.com/coveo/ui-kit/issues/5274)
+- chore(atomic): migrate FocusTargetController class (#5300) ([1b5c2a8](https://github.com/coveo/ui-kit/commits/1b5c2a8)), closes [#5300](https://github.com/coveo/ui-kit/issues/5300)
+- chore(atomic): migrate instant-item.tsx (#5349) ([7666f0c](https://github.com/coveo/ui-kit/commits/7666f0c)), closes [#5349](https://github.com/coveo/ui-kit/issues/5349)
+- chore(atomic): remove VNode type from breadcrumb interface (#5356) ([4f62235](https://github.com/coveo/ui-kit/commits/4f62235)), closes [#5356](https://github.com/coveo/ui-kit/issues/5356)
+- chore(atomic): simplify build process (#5310) ([edc7f24](https://github.com/coveo/ui-kit/commits/edc7f24)), closes [#5310](https://github.com/coveo/ui-kit/issues/5310)
+- feat(headless): explicit null trackingId if not specified (#5316) ([b04a573](https://github.com/coveo/ui-kit/commits/b04a573)), closes [#5316](https://github.com/coveo/ui-kit/issues/5316)
+- docs(atomic): fix bad link in query-suggestions storybook (#5348) ([e0420ca](https://github.com/coveo/ui-kit/commits/e0420ca)), closes [#5348](https://github.com/coveo/ui-kit/issues/5348)
+- fix(atomic): add back version export everywhere (#5330) ([dc880bb](https://github.com/coveo/ui-kit/commits/dc880bb)), closes [#5330](https://github.com/coveo/ui-kit/issues/5330)
+- fix(atomic): correctly localize tab label and don't localize tab name (#5339) ([6653341](https://github.com/coveo/ui-kit/commits/6653341)), closes [#5339](https://github.com/coveo/ui-kit/issues/5339)
+
+## <small>3.24.3 (2025-05-21)</small>
+
+- chore(atomic): migrate atomic-commerce-search-box-query-suggestions (#5205) ([1c9df4d](https://github.com/coveo/ui-kit/commits/1c9df4d)), closes [#5205](https://github.com/coveo/ui-kit/issues/5205)
+- chore(atomic): migrate recent-queries.tsx (#5255) ([7af6692](https://github.com/coveo/ui-kit/commits/7af6692)), closes [#5255](https://github.com/coveo/ui-kit/issues/5255)
+- fix(atomic): reduce scope of dispatchSearchBoxSuggestionsEvent to only relevant search boxes (#5253) ([d78ba42](https://github.com/coveo/ui-kit/commits/d78ba42)), closes [#5253](https://github.com/coveo/ui-kit/issues/5253) [#5205](https://github.com/coveo/ui-kit/issues/5205)
+
+## <small>3.24.2 (2025-05-21)</small>
+
+- chore(atomic): add tests for ItemListCommon class (#5278) ([55e3697](https://github.com/coveo/ui-kit/commits/55e3697)), closes [#5278](https://github.com/coveo/ui-kit/issues/5278)
+- fix(deps): update dependency typescript to v5.8.3 j:kit-282 (#5269) ([83d3475](https://github.com/coveo/ui-kit/commits/83d3475)), closes [#5269](https://github.com/coveo/ui-kit/issues/5269)
+
+## <small>3.24.1 (2025-05-15)</small>
+
+- fix(atomic): add missing import in atom (#5304) ([dac920d](https://github.com/coveo/ui-kit/commits/dac920d)), closes [#5304](https://github.com/coveo/ui-kit/issues/5304)
+- fix(atomic): ensure **all** custom-elements are scanned (#5302) ([fdeed93](https://github.com/coveo/ui-kit/commits/fdeed93)), closes [#5302](https://github.com/coveo/ui-kit/issues/5302)
+- fix(atomic): ensure cdn files includes a hash (#5306) ([fd03c71](https://github.com/coveo/ui-kit/commits/fd03c71)), closes [#5306](https://github.com/coveo/ui-kit/issues/5306) [#5305](https://github.com/coveo/ui-kit/issues/5305) [#5304](https://github.com/coveo/ui-kit/issues/5304) [#5303](https://github.com/coveo/ui-kit/issues/5303) [#5302](https://github.com/coveo/ui-kit/issues/5302)
+- fix(atomic): ensure context root is declared when context is used (#5305) ([89ff82e](https://github.com/coveo/ui-kit/commits/89ff82e)), closes [#5305](https://github.com/coveo/ui-kit/issues/5305)
+- fix(atomic): fix atomic-recs-result-template rendering error (#5313) ([7b4dd8a](https://github.com/coveo/ui-kit/commits/7b4dd8a)), closes [#5313](https://github.com/coveo/ui-kit/issues/5313) [/github.com/coveo/ui-kit/pull/5070/files#diff-869c29501cc46796f3035d2f9e0b3d564509824dd998c0c24d49efe4f2dd0423](https://github.com//github.com/coveo/ui-kit/pull/5070/files/issues/diff-869c29501cc46796f3035d2f9e0b3d564509824dd998c0c24d49efe4f2dd0423)
+- chore(atomic): add tests for ProductTemplateProvider class (#5276) ([e3734b4](https://github.com/coveo/ui-kit/commits/e3734b4)), closes [#5276](https://github.com/coveo/ui-kit/issues/5276)
+- chore(atomic): add tests for TemplateProvider class (#5275) ([10bc2a7](https://github.com/coveo/ui-kit/commits/10bc2a7)), closes [#5275](https://github.com/coveo/ui-kit/issues/5275)
+- chore(atomic): migrate display-grid.tsx (#5273) ([f0c3ed8](https://github.com/coveo/ui-kit/commits/f0c3ed8)), closes [#5273](https://github.com/coveo/ui-kit/issues/5273)
+- chore(atomic): migrate display-wrapper.tsx (#5266) ([a839b9c](https://github.com/coveo/ui-kit/commits/a839b9c)), closes [#5266](https://github.com/coveo/ui-kit/issues/5266)
+- chore(atomic): migrate placeholders.tsx (#5272) ([fe37143](https://github.com/coveo/ui-kit/commits/fe37143)), closes [#5272](https://github.com/coveo/ui-kit/issues/5272)
+- test(atomic): improve stability around suggestions/sb (#5303) ([99c7a3f](https://github.com/coveo/ui-kit/commits/99c7a3f)), closes [#5303](https://github.com/coveo/ui-kit/issues/5303)
+- docs(atomic): Storybook intro and crawling pages (#5293) ([57bb336](https://github.com/coveo/ui-kit/commits/57bb336)), closes [#5293](https://github.com/coveo/ui-kit/issues/5293)
+
 ## 3.24.0 (2025-05-08)
 
 - chore: add checkbox render function (#5215) ([fdb0da7](https://github.com/coveo/ui-kit/commits/fdb0da7)), closes [#5215](https://github.com/coveo/ui-kit/issues/5215)